refactor: add guards and rename
