+++ src/modules/scraper/service.ts
@@ -8,7 +8,7 @@ import { AppConfig } from "../configuration/configuration.service";
 import { ProcessedBlock } from "./model/ProcessedBlock.entity";
 import { MerkleDistributorProcessedBlock } from "./model/MerkleDistributorProcessedBlock.entity";
 import { ScraperQueuesService } from "./service/ScraperQueuesService";
-import { BlocksEventsQueueMessage, MerkleDistributorBlockEventsQueueMessage, ScraperQueue } from "./adapter/messaging";
+import { BlocksEventsQueueMessage, MerkleDistributorBlocksEventsQueueMessage, ScraperQueue } from "./adapter/messaging";
 import { wait } from "../../utils";
 
 @Injectable()
@@ -88,8 +88,8 @@ export class ScraperService {
     }
     await this.merkleDistributorProcessedBlockRepository.save(previousProcessedBlock);
 
-    await this.scraperQueuesService.publishMessage<MerkleDistributorBlockEventsQueueMessage>(
-      ScraperQueue.MerkleDistributorBlockEvents,
+    await this.scraperQueuesService.publishMessage<MerkleDistributorBlocksEventsQueueMessage>(
+      ScraperQueue.MerkleDistributorBlocksEvents,
       {
         chainId,
         ...blockRange,
